refactor(codex): extract encoding, revision, paths, and TOML leaves from prompt-layers (split S10 L1/2)
